add Mojito.Request.Single
