+++ test/mojito_test.exs
@@ -160,7 +160,7 @@ defmodule MojitoTest do
 
         assert(
           {:error, %{reason: :timeout}} =
-            Mojito.Request.receive_response(conn, %Mojito.Response{}, 100)
+            Mojito.Request.Single.receive_response(conn, %Mojito.Response{}, 100)
         )
       end
